refactor: component name modify
