+++ src/pages/flow-canvas/side-bar/Sidebar.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/mixins/scrollbar';
+@use '@/styles/mixins/scrollbar';
 
 .sidebarContainer {
   display: flex;
